checkpoint
Stream test with ffmpeg.wasm (front-end) and JavaCV (back-end) together.
Grabbing frames of a video file and encoding them in a separate video file which is directly sent as a response to the client. The client then grabs the frames from the MP4 and buffers them as images, and those images are played (or swapped one after another) at the desired speed.