Added a lot of custom exceptions
Need to make Spring return the corresponding response code of the exceptions as well as their messages, and not 401 Unauthorized with message "Full authentication is required bla bla bla"!

Also we need a request interceptor to check the User-Agent and eventually ban the user if their platform or browser is not the same as the registered one!

Lastly we need a logout endpoint, but NOT through a rest controller, but through Spring Security's logout function. This way the logout will be available for banned users too!